Using the three-part persistent project hooks
