+++ src/ProjectPageRouting.tsx
@@ -1,19 +1,21 @@
 import { ReactNode, useEffect } from "react";
 import { useRouterState } from "./router-hooks";
 import ProjectBrowser from "./project/ProjectBrowser";
-import { useProjectStorage } from "./project-persistence/ProjectStorageProvider";
+import { useProjectList } from "./project-persistence/project-list-hooks";
+import { usePersistentProject } from "./project-persistence/persistent-project-hooks";
 
 interface ProjectPageRoutingProps {
   children: ReactNode;
 }
 const ProjectPageRouting = ({ children }: ProjectPageRoutingProps) => {
   const [{ tab }] = useRouterState();
-  const { projectId, restoreMostRecentProject } = useProjectStorage();
+  const { projectList, restoreStoredProject } = useProjectList();
+  const { projectId } = usePersistentProject();
 
   useEffect(() => {
-    if (!projectId) {
+    if (!projectId && projectList) {
       const restoreState = async () => {
-        const restoredProject = await restoreMostRecentProject();
+        const restoredProject = await restoreStoredProject(projectList[0].id);
         if (!restoredProject && typeof tab !== "undefined") {
           history.replaceState(null, "", "/");
           window.dispatchEvent(new PopStateEvent("popstate"));
@@ -21,7 +23,7 @@ const ProjectPageRouting = ({ children }: ProjectPageRoutingProps) => {
       };
       void restoreState();
     }
-  }, []);
+  }, [projectId, projectList, restoreStoredProject, tab]);
 
   if (typeof tab === "undefined") {
     return <ProjectBrowser />;
