+++ src/project/project-hooks.tsx
@@ -22,7 +22,8 @@ import { useSessionSettings } from "../settings/session-settings";
 import { useSettings } from "../settings/settings";
 import { useSelection } from "../workbench/use-selection";
 import { defaultedProject, ProjectActions } from "./project-actions";
-import { useProjectStorage } from "../project-persistence/ProjectStorageProvider";
+import { useProjectList } from "../project-persistence/project-list-hooks";
+import { usePersistentProject } from "../project-persistence/persistent-project-hooks";
 
 /**
  * Hook exposing the main UI actions.
@@ -38,13 +39,10 @@ export const useProjectActions = (): ProjectActions => {
   const client = useLanguageServerClient();
   const [settings, setSettings] = useSettings();
   const [sessionSettings, setSessionSettings] = useSessionSettings();
-  const {
-    newStoredProject,
-    projectId,
-    restoreStoredProject,
-    setProjectName,
-    ydoc,
-  } = useProjectStorage();
+  const { projectId, ydoc } = usePersistentProject();
+
+  const { newStoredProject, restoreStoredProject, setProjectName } =
+    useProjectList();
   const actions = useMemo<ProjectActions>(
     () =>
       new ProjectActions(
@@ -58,8 +56,8 @@ export const useProjectActions = (): ProjectActions => {
         intl,
         logging,
         client,
-        ydoc,
-        projectId,
+        ydoc || null,
+        projectId || null,
         newStoredProject,
         restoreStoredProject,
         setProjectName
@@ -77,8 +75,10 @@ export const useProjectActions = (): ProjectActions => {
       setSettings,
       sessionSettings,
       setSessionSettings,
+      projectId,
       ydoc,
       newStoredProject,
+      restoreStoredProject,
       setProjectName,
     ]
   );
